CodecLZO

Documentation for CodecLZO.

CodecLZO.AbstractCommand

CodecLZO.HashMap

CodecLZO.HistoryCopyCommand

CodecLZO.LZO1X1CompressorCodec

CodecLZO.LZO1X1DecompressorCodec

CodecLZO.LZO1X1FastCompressorCodec

CodecLZO.LZO1X1FastDecompressorCodec

CodecLZO.LiteralCopyCommand

CodecLZO.ModuloBuffer

CodecLZO.PassThroughFIFO

Base.resize!

CodecLZO.capacity

CodecLZO.command_length

CodecLZO.compute_run_remainder

CodecLZO.copy_length

CodecLZO.count_matching

CodecLZO.decode_run

CodecLZO.encode_run!

CodecLZO.flush!

CodecLZO.lzo_compress

CodecLZO.lzo_compress!

CodecLZO.lzo_decompress

CodecLZO.lzo_decompress!

CodecLZO.multiplicative_hash

CodecLZO.pushout!

CodecLZO.reinterpret_get

CodecLZO.reinterpret_next

CodecLZO.repeatout!

CodecLZO.replace_all_matching!

CodecLZO.resize_front!

CodecLZO.shift_copy!

CodecLZO.state

CodecLZO.unsafe_decode_run

CodecLZO.unsafe_encode_run!

CodecLZO.unsafe_lzo_compress!

CodecLZO.unsafe_lzo_decompress!

CodecLZO.AbstractCommand — Type

AbstractCommand

A type representing either a literal copy (LiteralCopyCommand) or a history lookback copy (HistoryCopyCommand).

Types that inherit from AbstractCommand must implement the following methods:

command_length(::AbstractCommand)::Int, which returns the length of the encoded command in bytes;

copy_length(::AbstractCommand)::Int, which returns the number of bytes to be copied to the output;

one or both of:

decode(::Type{T}, ::AbstractVector{UInt8})::T where {T <: AbstractCommand}, which decodes a command of type T from the start of an AbstractVector{UInt8};

unsafe_decode(::Type{T}, ::Ptr{UInt8}, ::Integer)::T where {T <: AbstractCommand}, which decodes a command of type T from the memory pointed to by the pointer at a given (one-indexed) offset;

one or both of:

encode!(::T, ::AbstractCommand)::T where {T <: AbstractVector{UInt8}}, which encodes the command to the given vector and returns the modified vector;

unsafe_encode!(::Ptr{UInt8}, ::AbstractCommand, ::Integer)::Int, which encodes the command to the memory pointed to at a given (one-indexed offset) and returns the number of bytes written.

CodecLZO.HashMap — Type

HashMap{K,V}

A super-fast dictionary-like hash table of fixed size for integer keys.

CodecLZO.HistoryCopyCommand — Type

struct HistoryCopyCommand <: AbstractCommand

An encoded command representing a copy of a number of bytes from the already produced output back to the output.

In LZO1X, history lookback copies come in five varieties, the format of which is determined by the number of bytes copied, the lookback distance, and whether or not the previous command had a short literal copy tagged on the end:

Very short copy, short distance

Copies of three to four bytes with a lookback distance within 2048 bytes are encoded as two bytes, with bits encoding the length and distance. The command is to be interpreted in the following way (MSB first):

`01LDDDSS HHHHHHHH`

This means copy 3 + 0bL from a distance of 0b00000HHH_HHHHHDDD + 1.

The last two bits of the MSB instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Short copy, short distance

Copies of five to eight bytes with a lookback distance within 2048 bytes are encoded as two bytes, with bits encoding the length and distance. The command is to be interpreted in the following way (MSB first):

`1LLDDDSS HHHHHHHH`

This means copy 5 + 0bLL bytes from a distance of 0b00000HHH_HHHHHDDD + 1.

The last two bits of the MSB instruct the decoder to copy 0 through 3 literals from the input to the output immediately following the history lookback copy.

Any length copy, short to medium distance

Copies of any length greater than two with a lookback distance within 16384 bytes is incoded with at least three bytes and with as many as necessary to encode the run length. The command is to be interpreted in the following way (MSB first):

`001LLLLL [Z zero bytes] [XXXXXXXX] EEEEEESS DDDDDDDD`

The lower five bits of the first byte represent the length of the copy minus two. This can obviously only represent copies of length 2 to 33, so to encode longer copies, LZO1X uses the following encoding method:

If 0bLLLLL is non-zero, then length = 2 + 0bLLLLL

If 0bLLLLL is zero, then length = 33 + (number of zero bytes after the first) × 255 + (first non-zero byte)

The lookback distance is encoded in LE order in the last two bytes: that is, the last byte of the command holds the MSB of the distance, and the second-to-last byte holds the LSB of the distance. The distance is interpreted as distance = 0b00DDDDDD_DDEEEEEE + 1.

The last two bits of the second-to-last byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Any length copy, long distance

Copies of any length greater than two with a lookback distance between 16385 and 49151 bytes is incoded with at least three bytes and with as many as necessary to encode the run length. The command is to be interpreted in the following way (MSB first):

`0001HLLL [Z zero bytes] [XXXXXXXX] EEEEEESS DDDDDDDD`

As with other variable length runs,, LZO1X uses the following encoding method with this command:

If 0bLLL is non-zero, then length = 2 + 0bLLL

If 0bLLL is zero, then length = 9 + (number of zero bytes after the first) × 255 + (first non-zero byte)

The lookback distance is encoded with one bit in the first command byte (H), then in LE order in the last two bytes: that is, the last byte of the command holds the MSB of the distance, and the second-to-last byte holds the LSB of the distance. The distance is interpreted as distance = 16384 + 0b0HDDDDDD_DDEEEEEE.

The last two bits of the second-to-last byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Special End-of-Stream Encoding

End-of-stream is signaled by a history lookback copy of 3 bytes from a distance of 16384 bytes with no subsequent literal copies. This corresponds to a long historical lookback copy command of 0b00010001 0b00000000 0b00000000.

Note

The maximum lookback distance that LZO1X can encode is 16384 + 0b01111111_11111111 == 49151 bytes.

Short copies from short distances following literal copies

If the previous history lookback command included a short literal copy (1 ≤ 0bSS ≤ 3, as encoded in the above commands), then a special two-byte command can be used to copy two bytes with a lookback distance within 1024 bytes. The command is to be interpreted in the following way (MSB first):

`0000DDSS HHHHHHHH`

The number of bytes to copy is always length = 2, and the lookback distance is 0b000000HH_HHHHHHDD + 1.

The last two bits of the first byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

If the previous command was a long literal copy (of four of more bytes), then the same two-byte command means something different: it encodes a three-byte copy with a lookback distance between 2049 and 3071 bytes. The command is interpreted the same as above, but length = 3 and distance = 0b000000HH_HHHHHHDD + 2049.

Note

Because encoding these special commands require a historical match of fewer than four bytes, they are never encoded by the LZO1X algorithm: however, they are valid LZO1X commands, and the LZO1X decoder will interpret them correctly.

See also CodecLZO.LiteralCopyCommand.

CodecLZO.LZO1X1CompressorCodec — Type

LZO1X1CompressorCodec(level::Int=5) <: TranscodingStreams.Codec

A struct that compresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<12 = 4096 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation needs to keep 49151 bytes of input history in memory in addition to the 4096-byte hash map, but only expands the output as necessary during compression.

Arguments:

level::Int = 5: The speed/compression tradeoff paramter, with larger numbers representing slower compression at a higher compresison ratio. On a technical level, every pow(2, level) history misses increases by 1 the number of bytes skipped when searching for the next history match. The default (5) is recommended by the original liblzo2 authors as a good balance between speed and compression.

CodecLZO.LZO1X1DecompressorCodec — Type

LZO1X1DecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

For implementation purposes, this decompressor uses a buffer of 49151 bytes to store output. This is equal to the maximum lookback distance of the LZO 1X1 algorithm.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation needs to keep 49151 bytes of output history in memory while decompressing, equal to the maximum lookback distance of the LZO 1x1 algorithm, and a small number of bytes to keep track of the command being processed in case the command is broken between multiple reads from the input memory.

CodecLZO.LZO1X1FastCompressorCodec — Type

LZO1X1FastCompressorCodec <: TranscodingStreams.Codec

A struct that compresses data using the liblzo2 version version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<16 = 65536 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation uses an expanding input buffer that waits until all input is available before processing, eliminating the usefulness of the TranscodingStreams interface.

CodecLZO.LZO1X1FastDecompressorCodec — Type

LZO1X1FastDecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data using liblzo1 library version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and therefore does not take advantage of the memory savings allowed by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation reports a very large memory requirement with TranscodingStreams.minoutsize to account for this.

CodecLZO.LiteralCopyCommand — Type

struct LiteralCopyCommand <: AbstractCommand

An encoded command representing a copy of a number of bytes from input straight to output.

In LZO1X, literal copies come in three varieties:

Long copies

LZO1X long copy commands begin with a byte with four high zero bits and four low potentially non-zero bits:

0 0 0 0 L L L L

The low four bits represent the length of the copy minus three. This can obviously only represent copies of length 3 to 18, so to encode longer copies, LZO1X uses the following encoding method:

If the first byte is non-zero, then length = 3 + L

If the first byte is zero, then length = 18 + (number of zero bytes after the first) × 255 + (first non-zero byte)

This means a length of 18 is encoded as [0b00001111], a length of 19 is encoded as [0b00000000, 0b00000001], a length of 274 is encoded as [0b00000000, 0b00000000, 0b00000001], and so on.

Short copies

The long copy command cannot encode copies shorter than four bytes by design. If a literal of three or fewer bytes needs to be copied, it is encoded in the two least significant bits of the previous history lookback copy command. This works because literal copies and history lookback copies always alternate in LZO1X streams.

First literal copies

LZO1X streams always begin with a literal copy command of at least four bytes. Because the first command is always a literal copy, a special format is used to copy runs of literals that are between 18 and 238 bytes that compacts the command into a single byte. If the first byte of the stream has the following values, they are interpreted as the corresponding literal copy commands:

0:15: Treat as a "long copy" encoding (see above).

17:255: Treat as a copy of (byte - 17) literals.

Note that 17:20 are invalid values for a first copy command in LZO1X streams because history lookback copy lengths must always be four or more bytes. A value of 16 in the first position is always invalid.

Note

The official liblzo2 version of LZO1X properly decodes these first literal copy codes, but never encodes them when compressing data.

See also CodecLZO.HistoryCopyCommand.

CodecLZO.ModuloBuffer — Type

ModuloBuffer{T}(n::Integer)
ModuloBuffer(iter)

An AbstractVector{T} of fixed capacity n with periodic boundary conditions on the index.

The first version of the constructor will create an empty buffer with a capacity of n. The second version will copy all elements of the iterable object iter into a new buffer with a capacity of length(iter) and element type eltype(iter).

If a new element added (either with push!, pushfirst!, append!, or prepend!) would increase the size of the buffer past the capacity, the oldest element added will be overwritten (or the newest element added in the case of pushfirst! or prepend!) to maintain the fixed capacity.

If the buffer is not at capacity, then attempts to index into unfilled elements of the buffer will result in a BoundsError.

CodecLZO.PassThroughFIFO — Type

PassThroughFIFO <: AbstractVector{UInt8}

A FIFO (first in, first out data structure) that buffers data pushed into the front of it and, when full, pushes out older data from the back to a sink.

Base.resize! — Method

resize!(buffer::ModuloBuffer, n::Integer)::ModuloBuffer

Resize buffer to a capacity of n elements by adjusting the location of the back of the buffer.

If n < capacity(buffer), only the first n elements of buffer will be retained.

Attempts to avoid allocating new memory by manipulating and resizing the internal vector in place.

CodecLZO.capacity — Method

capacity(buffer::ModuloBuffer)::Int

Return the maximum number of elements buffer can contain.

CodecLZO.command_length — Method

command_length(command)::Int

Return the number of bytes in the encoded command.

CodecLZO.compute_run_remainder — Method

compute_run_remainder(n, bits)::Tuple{Int, Int}

Compute the number of bytes necessary to encode a run of length n given a first-byte mask of length bits, also returning the remainder byte.

Note

This method does not adjust n before computing the run length. Perform adjustments before calling this method.

CodecLZO.copy_length — Method

copy_length(command)::Int

Return the number of bytes that are to be copied to the output by command.

CodecLZO.count_matching — Method

count_matching(a::AbstractVector, b::AbstractVector)

Count the number of elements at the start of a that match the elements at the start of b.

Equivalent to findfirst(a .!= b), but faster and limiting itself to the first min(length(a), length(b)) elements.

CodecLZO.decode_run — Method

decode_run(input::Vector{UInt8}, bits)::Tuple{Int, Int}

Decode the length of the run in bytes and the number of bytes to copy from input given a mask of bits bits.

CodecLZO.encode_run! — Method

encode_run!(output, len, bits)::Int

Emit the number of zero bytes necessary to encode a length len in a command expecting bits leading bits, returning the number of bytes written to the output.

Literal and copy lengths are always encoded as either a single byte or a sequence of three or more bytes. If len < (1 << bits), the length will be encoded in the lower bits bits of the starting byte of output so the return will be 0. Otherwise, the return will be the number of additional bytes needed to encode the length. The returned number of bytes does not include the zeros in the first byte (the command) used to signal that a run encoding follows, but it does include the remainder.

Note: the argument len is expected to be the adjusted length for the command. Literals use an adjusted length of len = length(literal) - 3 and copy commands use an adjusted literal length of len = length(copy) - 2.

CodecLZO.flush! — Method

flush!(p::PassThroughFIFO, sink::AbstractVector{UInt8})

Copy all the data in p to the front of sink.

Returns the number of bytes copied, equal to min(length(p), length(sink)). If length(sink) >= length(p), isempty(p) == true after the flush, else length(p) will be equal to the number of bytes that could not be pushed to sink.

CodecLZO.lzo_compress — Function

lzo_compress(src, [working_memory=zeros(UInt8, 1<<12)])::Vector{UInt8}

Compress src using the LZO 1X1 algorithm.

Returns a compressed version of src.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_compress! — Function

lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])

Compress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the compressed data if necessary. Returns the modified dest.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_decompress! — Method

lzo_decompress!(dest::Vector{UInt8}, src)

Decompress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the decompressed data if necessary. Returns the modified dest.

CodecLZO.lzo_decompress — Method

lzo_decompress(src)::Vector{UInt8}

Decompress src using the LZO 1X1 algorithm.

Returns a decompressed version of src.

CodecLZO.multiplicative_hash — Method

multiplicative_hash(value, magic_number, bits, [mask::V = typemax(UInt64)])

Hash value into a type V using multiplicative hashing.

This method performs floor((value * magic_number % W) / (W / M)) where W = 2^64, M = 2^m, and magic_number is relatively prime to W, is large, and has a good mix of 1s and 0s in its binary representation. In modulo 2^64 arithmetic, this becomes (value * magic_number) >>> m.

CodecLZO.pushout! — Method

pushout!(p::PassThroughFIFO, source, sink::AbstractVector{UInt8})

Push as much of source into the FIFO as it can hold, pushing out stored data to sink.

The argument source can be an AbstractVector{UInt8} or a single UInt8 value.

Until p is full, elements from source will be added to the FIFO and no elements will be pushed out to sink. Once p is full, elements of source up to capacity(p) will be added to the FIFO and the older elements will be pushed to sink.

Returns a tuple of the number of elements read from source and the number of elements written to sink.

See repeatout!.

CodecLZO.reinterpret_get — Method

reinterpret_get(T::Type, input, [index::Integer = 1])::T

Reinterpret bytes from input as an LE-ordered value of type T, optionally starting at index. This tries to be faster than reinterpret(T, input[index:index+sizeof(T)-1]).

input can be anything that is linearly indexed and getindex(input, ::Int) returns a type S for which the operator |(::T, ::S) is defined.

CodecLZO.reinterpret_next — Method

reinterpret_next(previous::T, input::AbstractVector{UInt8}, [index::Int = 1])::T

Get the byte from input at index and push it to the LSB of previous, rotating off the MSB. This tries to be faster than doing reinterpret(T, input[index:index+sizeof(T)-1]) twice by reusing the already read LSBs.

CodecLZO.repeatout! — Method

repeatout!(p::PassThroughFIFO, lookback::Integer, n::Integer, sink::AbstractVector{UInt8})

Append n values starting from lookback bytes from the end of p to the front of p.

Once p is full, any bytes that are appended to the front of p will cause bytes from the back to be expelled into the front of sink.

This method works even if n > lookback, in which case the bytes that were appended to the front of p first will be repeated.

Returns the number of bytes expelled from p into sink.

CodecLZO.replace_all_matching! — Method

replace_all_matching!(h::HashMap, input, input_start, output, output_start)

Count the number of elements at the start of input that match the elements at the start of output, putting the matching indices of input as values into h keyed by the K integer read from input at that index.

Returns the number of matching bytes found (not necessarily equal to the number of Ks put into h).

CodecLZO.resize_front! — Method

resize_front!(buffer::ModuloBuffer, n::Integer)::ModuloBuffer

Resize buffer to a capacity of n elements by adjusting the location of the front of the buffer.

If n < capacity(buffer), only the last n elements of buffer will be retained.

Attempts to avoid allocating new memory by manipulating and resizing the internal vector in place.

CodecLZO.shift_copy! — Function

shift_copy!(buffer::ModuloBuffer, source, i, sink, j, [n])::Tuple{Int,Int}

Copy n elements source starting at i to the back of buffer, evicting elements from the front of buffer to sink starting at j.

source must be a valid source for the copyto! method, and sink must be a valid destination.

If not specified, n defaults to the minimum of the capacity of buffer, the number of elements that can be copied from source, and the amount of free space left in sink.

If n <= capacity(buffer) - length(buffer), then all n elements will be copied into the available space in buffer and no elements will be evicted from the front of buffer into sink.

Returns a tuple of the number of elements copied from source and the number of elements evicted from buffer and copied to sink.

CodecLZO.state — Method

state(codec)::MatchingState

Determine the state of the codec from the command in the buffer.

The state of the codec can be one of:

FIRST_LITERAL`: in the middle of recording the first literal copy command from the input (the initial state);

LITERAL: in the middle of writing a literal copy command to the output; or

HISTORY: in the middle of writing a history copy command to the output.

CodecLZO.unsafe_decode_run — Method

unsafe_decode_run(p::Ptr{UInt8}, i, bits)::Tuple{Int, Int}

Decode the length of the run in bytes and the number of bytes to copy from the memory address pointed to by p offset by i given a mask of bits bits.

This method is "unsafe" in that it will not stop reading from memory addresses after p until it finds a non-zero byte, whatever the consequences.

CodecLZO.unsafe_encode_run! — Function

unsafe_encode_run!(p::Ptr{UInt8}, len, bits, [i=1])::Int

Emit the number of zero bytes necessary to encode a length len in a command expecting bits leading bits, returning the number of bytes written to the output.

Literal and copy lengths are always encoded as either a single byte or a sequence of three or more bytes. If len < (1 << bits), the length will be encoded in the lower bits bits of the starting byte of output so the return will be 0. Otherwise, the return will be the number of additional bytes needed to encode the length. The returned number of bytes does not include the zeros in the first byte (the command) used to signal that a run encoding follows, but it does include the remainder.

This method is "unsafe" in that it does not check if p points to an area of memory large enough to hold the resulting run before clobbering it.

Note: the argument len is expected to be the adjusted length for the command. Literals use an adjusted length of len = length(literal) - 3 and copy commands use an adjusted literal length of len = length(copy) - 2.

CodecLZO.unsafe_lzo_compress! — Function

unsafe_lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])::Int

Compress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the compressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to compress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.unsafe_lzo_decompress! — Method

unsafe_lzo_decompress!(dest::Vector{UInt8}, src)::Int

Decompress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the decompressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to decompress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).