CodecLZO

Documentation for CodecLZO.

CodecLZO.HashMap

CodecLZO.LZO1X1CompressorCodec

CodecLZO.count_matching

CodecLZO.encode_run!

CodecLZO.multiplicative_hash

CodecLZO.reinterpret_get

CodecLZO.HashMap — Type

HashMap{K,V}

A super-fast dictionary-like hash table of fixed size for integer keys.

CodecLZO.LZO1X1CompressorCodec — Type

LZO1X1CompressorCodec <: AbstractLZOCompressorCodec

A TranscodingStreams.Codec struct that compresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<12 = 4096 elements;

A 4-byte history lookup window that scans the input with a logarithmically increasing skip distance;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation needs to keep 49151 bytes of input history in memory in addition to the 4096-byte hash map, but only expands the output as necessary during compression.

CodecLZO.count_matching — Method

count_matching(a::AbstractVector, b::AbstractVector)

Count the number of elements at the start of a that match the elements at the start of b.

Equivalent to findfirst(a .!= b), but faster and limiting itself to the first min(length(a), length(b)) elements.

CodecLZO.encode_run! — Method

n_written = encode_run(output, start_index, len, bits)

Emit the number of zero bytes necessary to encode a length len in a command expecting bits leading bits.

Literal and copy lengths are always encoded as either a single byte or a sequence of three or more bytes. If len < (1 << bits), the length will be encoded in the lower bits bits of the starting byte of output so the return will be 0. Otherwise, the return will be the number of additional bytes needed to encode the length. The returned number of bytes does not include the zeros in the first byte (the command) used to signal that a run encoding follows, but it does include the remainder.

Note: the argument len is expected to be the adjusted length for the command. Literals use an adjusted length of len = length(literal) - 3 and copy commands use an adjusted literal length of len = length(copy) - 2.

CodecLZO.multiplicative_hash — Method

multiplicative_hash(value, magic_number, bits, [mask::V = typemax(UInt64)])

Hash value into a type V using multiplicative hashing.

This method performs floor((value * magic_number % W) / (W / M)) where W = 2^64, M = 2^m, and magic_number is relatively prime to W, is large, and has a good mix of 1s and 0s in its binary representation. In modulo 2^64 arithmetic, this becomes (value * magic_number) >>> m.

CodecLZO.reinterpret_get — Method

reinterpret_get(T::Type, input::AbstractVector{UInt8}, [index::Int = 1])::T

Reinterpret bytes from input as an LE-ordered value of type T, optionally starting at index. This tries to be faster than reinterpret(T, input[index:index+sizeof(T)-1]).