CodecLZO

Documentation for CodecLZO.

CodecLZO.HashMap

CodecLZO.LZO1X1CompressorCodec

CodecLZO.LZO1X1DecompressorCodec

CodecLZO.LZO1X1FastCompressorCodec

CodecLZO.LZO1X1FastDecompressorCodec

CodecLZO.PassThroughFIFO

CodecLZO.count_matching

CodecLZO.encode_run!

CodecLZO.lzo_compress

CodecLZO.lzo_compress!

CodecLZO.lzo_decompress

CodecLZO.lzo_decompress!

CodecLZO.multiplicative_hash

CodecLZO.prepend!

CodecLZO.reinterpret_get

CodecLZO.reinterpret_next

CodecLZO.replace_all_matching!

CodecLZO.unsafe_lzo_compress!

CodecLZO.unsafe_lzo_decompress!

CodecLZO.HashMap — Type

HashMap{K,V}

A super-fast dictionary-like hash table of fixed size for integer keys.

CodecLZO.LZO1X1CompressorCodec — Type

LZO1X1CompressorCodec <: TranscodingStreams.Codec

A struct that compresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<12 = 4096 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation needs to keep 49151 bytes of input history in memory in addition to the 4096-byte hash map, but only expands the output as necessary during compression.

CodecLZO.LZO1X1DecompressorCodec — Type

LZO1X1DecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

For implementation purposes, this decompressor uses a buffer of 49151 bytes to store output. This is equal to the maximum lookback distance of the LZO 1X1 algorithm.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation needs to keep 49151 bytes of output history in memory while decompressing, equal to the maximum lookback distance of the LZO 1x1 algorithm, and a small number of bytes to keep track of the command being processed in case the command is broken between multiple reads from the input memory.

CodecLZO.LZO1X1FastCompressorCodec — Type

LZO1X1FastCompressorCodec <: TranscodingStreams.Codec

A struct that compresses data using the liblzo2 version version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<16 = 65536 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation uses an expanding input buffer that waits until all input is available before processing, eliminating the usefulness of the TranscodingStreams interface.

CodecLZO.LZO1X1FastDecompressorCodec — Type

LZO1X1FastDecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data using liblzo1 library version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and therefore does not take advantage of the memory savings allowed by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation reports a very large memory requirement with TranscodingStreams.minoutsize to account for this.

CodecLZO.PassThroughFIFO — Type

PassThroughFIFO <: AbstractVector{UInt8}

A FIFO (first in, first out) that buffers data pushed into it and pushes out older data to a sink when new data is prepended.

CodecLZO.count_matching — Method

count_matching(a::AbstractVector, b::AbstractVector)

Count the number of elements at the start of a that match the elements at the start of b.

Equivalent to findfirst(a .!= b), but faster and limiting itself to the first min(length(a), length(b)) elements.

CodecLZO.encode_run! — Method

n_written = encode_run(output, start_index, len, bits)

Emit the number of zero bytes necessary to encode a length len in a command expecting bits leading bits.

Literal and copy lengths are always encoded as either a single byte or a sequence of three or more bytes. If len < (1 << bits), the length will be encoded in the lower bits bits of the starting byte of output so the return will be 0. Otherwise, the return will be the number of additional bytes needed to encode the length. The returned number of bytes does not include the zeros in the first byte (the command) used to signal that a run encoding follows, but it does include the remainder.

Note: the argument len is expected to be the adjusted length for the command. Literals use an adjusted length of len = length(literal) - 3 and copy commands use an adjusted literal length of len = length(copy) - 2.

CodecLZO.lzo_compress — Function

lzo_compress(src, [working_memory=zeros(UInt8, 1<<12)])::Vector{UInt8}

Compress src using the LZO 1X1 algorithm.

Returns a compressed version of src.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_compress! — Function

lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])

Compress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the compressed data if necessary. Returns the modified dest.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_decompress! — Method

lzo_decompress!(dest::Vector{UInt8}, src)

Decompress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the decompressed data if necessary. Returns the modified dest.

CodecLZO.lzo_decompress — Method

lzo_decompress(src)::Vector{UInt8}

Decompress src using the LZO 1X1 algorithm.

Returns a decompressed version of src.

CodecLZO.multiplicative_hash — Method

multiplicative_hash(value, magic_number, bits, [mask::V = typemax(UInt64)])

Hash value into a type V using multiplicative hashing.

This method performs floor((value * magic_number % W) / (W / M)) where W = 2^64, M = 2^m, and magic_number is relatively prime to W, is large, and has a good mix of 1s and 0s in its binary representation. In modulo 2^64 arithmetic, this becomes (value * magic_number) >>> m.

CodecLZO.prepend! — Method

prepend!(p::PassThroughFIFO, source, source_start, sink, sink_start)

Push as much of source (starting at source_start) into the FIFO as it can hold, pushing out stored data to sink (starting at sink_start).

Until p is full, elements from source will be added to the FIFO and no elements will be pushed out to sink. Once p is full, elements of source up to capacity(p) will be added to the FIFO and the older elements will be pushed to sink.

Returns a tuple of the number of elements read from source and the number of elements written to sink.

CodecLZO.reinterpret_get — Method

reinterpret_get(T::Type, input::AbstractVector{UInt8}, [index::Int = 1])::T

Reinterpret bytes from input as an LE-ordered value of type T, optionally starting at index. This tries to be faster than reinterpret(T, input[index:index+sizeof(T)-1]).

CodecLZO.reinterpret_next — Method

reinterpret_next(previous::T, input::AbstractVector{UInt8}, [index::Int = 1])::T

Get the byte from input at index and push it to the LSB of previous, rotating off the MSB. This tries to be faster than doing reinterpret(T, input[index:index+sizeof(T)-1]) twice by reusing the already read LSBs.

CodecLZO.replace_all_matching! — Method

replace_all_matching!(h::HashMap, input, input_start, output, output_start)

Count the number of elements at the start of input that match the elements at the start of output, putting the matching indices of input as values into h keyed by the K integer read from input at that index.

Returns the number of matching bytes found (not necessarily equal to the number of Ks put into h).

CodecLZO.unsafe_lzo_compress! — Function

unsafe_lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])::Int

Compress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the compressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to compress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.unsafe_lzo_decompress! — Method

unsafe_lzo_decompress!(dest::Vector{UInt8}, src)::Int

Decompress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the decompressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to decompress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).