CodecLZO

Documentation for CodecLZO.

CodecLZO.CircularBuffer

CodecLZO.LZO1X1CompressorCodec

CodecLZO.CircularBuffer — Type

CircularBuffer{T} <: AbstractVector{T}

A buffer for simultaneous writing to and reading from a contiguous array of memory with periodic boundries.

CodecLZO.LZO1X1CompressorCodec — Type

LZO1X1CompressorCodec <: AbstractLZOCompressorCodec

A TranscodingStreams.Codec struct that compresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is defined by:

A lookback dictionary implemented as a hash map with a maximum of size of at most 1<<12 = 4096 elements and as few as 16 elements using a specific fast hashing algorithm;

An 8-byte history lookup window that scans the input with a logarithmically increasing skip distance;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version therefore uses only a 4096-byte hash map as additional working memory, while this version needs to keep the full 49151 bytes of history in memory in addition to the 4096-byte hash map.