CodecLZO

Documentation for CodecLZO.

CodecLZO.CommandPair

CodecLZO.HashMap

CodecLZO.LZO1X1CompressorCodec

CodecLZO.LZO1X1DecompressorCodec

CodecLZO.LZO1X1FastCompressorCodec

CodecLZO.LZO1X1FastDecompressorCodec

Base.copyto!

Base.copyto!

Base.copyto!

Base.copyto!

CodecLZO.command_length

CodecLZO.compute_run_remainder

CodecLZO.decode

CodecLZO.decode_run

CodecLZO.encode!

CodecLZO.encode_run

CodecLZO.encode_run!

CodecLZO.lzo_compress

CodecLZO.lzo_compress!

CodecLZO.lzo_decompress

CodecLZO.lzo_decompress!

CodecLZO.match_length

CodecLZO.maxcopy!

CodecLZO.multiplicative_hash

CodecLZO.reinterpret_get

CodecLZO.reinterpret_next

CodecLZO.replace_all_matching!

CodecLZO.state

CodecLZO.state

CodecLZO.unsafe_lzo_compress!

CodecLZO.unsafe_lzo_decompress!

CodecLZO.CommandPair — Type

CommandPair

A mutable type representing both a history lookback copy and a literal copy.

Except for the first or last command, commands in LZO 1X1 always come in pairs: a history lookback always follows a literal, and if a literal copy of zero bytes is considered, a literal copy always follows a history lookback. This means commands can be efficiently stored, parsed, and encoded as pairs of commands.

Literal copies

In LZO1X, literal copies come in three varieties:

Long copies

LZO1X long copy commands begin with a byte with four high zero bits and four low potentially non-zero bits:

`0000LLLL [Z zero bytes] [XXXXXXXX]`

The low four bits represent the length of the copy minus three. This can obviously only represent copies of length 3 to 18, so to encode longer copies, LZO1X uses the following encoding method:

If the first byte is non-zero, then length = 3 + L

If the first byte is zero, then length = 18 + Z × 255 + 0bXXXXXXXX

This means a length of 18 is encoded as [0b00001111], a length of 19 is encoded as [0b00000000, 0b00000001], a length of 274 is encoded as [0b00000000, 0b00000000, 0b00000001], and so on.

Short copies

The long copy command cannot encode copies shorter than four bytes by design. If a literal of three or fewer bytes needs to be copied, it is encoded in the two least significant bits of the previous history lookback copy command. This works because literal copies and history lookback copies always alternate in LZO1X streams.

First literal copies

LZO1X streams always begin with a literal copy command of at least four bytes. Because the first command is always a literal copy, a special format is used to copy runs of literals that are between 18 and 238 bytes that compacts the command into a single byte. If the first byte of the stream has the following values, they are interpreted as the corresponding literal copy commands:

0:15: Treat as a "long copy" encoding (see above).

17:255: Treat as a copy of (byte - 17) literals.

Note

The first literal copy command of 17 is technically a valid first copy command, even though it signals a copy of zero bytes; however, if the first byte of an LZO stream is 17, the very next bytes must encode a long literal copy because there is no history available for a history copy command, and LZO does not have a way to encode a zero-byte history copy. A value of 16 in the first position is always invalid because it signals a history lookback copy command which cannot come before any literals are copied to the output.

History lookback copies

In LZO1X, history lookback copies come in five varieties, the format of which is determined by the number of bytes copied, the lookback distance, and whether or not the previous command had a short literal copy tagged on the end:

Short copy, short distance

Copies of five to eight bytes with a lookback distance within 2048 bytes are encoded as two bytes, with bits encoding the length and distance. The command is to be interpreted in the following way (MSB first):

`1LLDDDSS HHHHHHHH`

This means copy 5 + 0bLL bytes from a distance of 0b00000HHH_HHHHHDDD + 1.

The last two bits of the MSB instruct the decoder to copy 0 through 3 literals from the input to the output immediately following the history lookback copy.

Very short copy, short distance

Copies of three to four bytes with a lookback distance within 2048 bytes are encoded as two bytes, with bits encoding the length and distance. The command is to be interpreted in the following way (MSB first):

`01LDDDSS HHHHHHHH`

This means copy 3 + 0bL from a distance of 0b00000HHH_HHHHHDDD + 1.

The last two bits of the MSB instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Note

LZO1X checks for historical matches based on four byte runs, so a historical copy of three bytes can only occur if the byte that follows also matches, in which case the run is at least four bytes long. The reference LZO1X algorithm used in liblzo2 is a greedy algorithm, so it will only ever encode a three byte historical copy in the special case where the first three bytes of a sequence are repeated exactly once and are followed by a non-matching byte.

Any length copy, short to medium distance

Copies of any length greater than two with a lookback distance within 16384 bytes are incoded with at least three bytes and with as many as necessary to encode the run length. The command is to be interpreted in the following way (MSB first):

`001LLLLL [Z zero bytes] [XXXXXXXX] EEEEEESS DDDDDDDD`

The lower five bits of the first byte represent the length of the copy minus two. To encode copies longer than 33 bytes, LZO1X uses the following encoding method:

If 0bLLLLL is non-zero, then length = 2 + 0bLLLLL

If 0bLLLLL is zero, then length = 33 + Z × 255 + 0bXXXXXXXX

The lookback distance is encoded in LE order in the last two bytes: that is, the last byte of the command holds the MSB of the distance, and the second-to-last byte holds the LSB of the distance. The distance is interpreted as distance = 0b00DDDDDD_DDEEEEEE + 1.

The last two bits of the second-to-last byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Any length copy, long distance

Copies of any length greater than two with a lookback distance between 16385 and 49151 bytes are incoded with at least three bytes and with as many as necessary to encode the run length. The command is to be interpreted in the following way (MSB first):

`0001HLLL [Z zero bytes] [XXXXXXXX] EEEEEESS DDDDDDDD`

As with other variable length runs, LZO1X uses the following encoding method with this command:

If 0bLLL is non-zero, then length = 2 + 0bLLL

If 0bLLL is zero, then length = 9 + Z × 255 + 0bXXXXXXXX

The lookback distance is encoded with one bit in the first command byte (H), then in LE order in the last two bytes: that is, the last byte of the command holds the MSB of the distance, and the second-to-last byte holds the LSB of the distance. The distance is interpreted as distance = 16384 + 0b0HDDDDDD_DDEEEEEE.

The last two bits of the second-to-last byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

Special End-of-Stream Encoding

End-of-stream is signaled by a history lookback copy of 3 bytes from a distance of 16384 bytes with no subsequent literal copies. This corresponds to two possible commands, so by convention the end-of-stream command uses the long distance version (0b00010001 0b00000000 0b00000000) while an actual copy of 3 bytes from a distance of 16384 uses the short distance version (0b00100001 0b00000000 0b00000000).

Note

The maximum lookback distance that LZO1X can encode is 16384 + 0b01111111_11111111 == 49151 bytes.

Short copies from short distances following literal copies

If the previous history lookback command included a short literal copy (1 ≤ 0bSS ≤ 3, as encoded in the above commands), then a special two-byte command can be used to copy two bytes with a lookback distance within 1024 bytes. The command is to be interpreted in the following way (MSB first):

`0000DDSS HHHHHHHH`

The number of bytes to copy is always length = 2, and the lookback distance is 0b000000HH_HHHHHHDD + 1.

The last two bits of the first byte instruct the decoder to copy 0bSS literals from the input to the output immediately following the history lookback copy.

If the previous command was a long literal copy (of four of more bytes), then the same two-byte command means something different: it encodes a three-byte copy with a lookback distance between 2049 and 3072 bytes. The command is interpreted the same as above, but length = 3 and distance = 0b000000HH_HHHHHHDD + 2049.

Note

LZO1X checks for historical matches based on four byte runs, so a historical copy of two bytes can only occur if the two bytes that follow also match, in which case the run is at least four bytes long. The reference LZO1X algorithm used in liblzo2 is a greedy algorithm, so it will only ever encode a two byte historical copy in the special case where the first two bytes of a sequence repeat exactly once and are followed by a non-matching byte.

See also: decode and encode!.

CodecLZO.HashMap — Type

HashMap{K,V}

A super-fast dictionary-like hash table of fixed size for integer keys.

CodecLZO.LZO1X1CompressorCodec — Type

LZO1X1CompressorCodec(level::Int=5) <: TranscodingStreams.Codec

A struct that compresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<12 = 4096 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation needs to keep 98310 bytes of input history in memory in addition to the 4096-byte hash map, and also caches literal copies in an array that expands as necessary during compression.

Arguments:

level::Int = 5: The speed/compression tradeoff paramter, with larger numbers representing slower compression at a higher compresison ratio. On a technical level, every pow(2, level) history misses increases by 1 the number of bytes skipped when searching for the next history match. The default (5) is recommended by the original liblzo2 authors as a good balance between speed and compression.

CodecLZO.LZO1X1DecompressorCodec — Type

LZO1X1DecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data according to the 1X1 version of the LZO algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

For implementation purposes, this decompressor uses a buffer of 49151 bytes to store output. This is equal to the maximum lookback distance of the LZO 1X1 algorithm.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and is therefore not adaptable to streaming as required by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation needs to keep 49151 bytes of output history in memory while decompressing, equal to the maximum lookback distance of the LZO 1x1 algorithm, and a small number of bytes to keep track of the command being processed in case the command is broken between multiple reads from the input memory.

CodecLZO.LZO1X1FastCompressorCodec — Type

LZO1X1FastCompressorCodec <: TranscodingStreams.Codec

A struct that compresses data using the liblzo2 version version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm defined by:

A lookback dictionary implemented as a hash map with a maximum of size of 1<<16 = 65536 elements;

A 4-byte history lookup window that scans the input with a skip distance that increases linearly with the number of misses;

A maximum lookback distance of 0b11000000_00000000 - 1 = 49151 bytes;

The C implementation of LZO defined by liblzo2 requires that all compressable information be loaded in working memory at once. The C library version claims to use only a 4096-byte hash map as additional working memory, but it also requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the uncompressed data by a factor of roughly 256/255. This implementation uses an expanding input buffer that waits until all input is available before processing, eliminating the usefulness of the TranscodingStreams interface.

CodecLZO.LZO1X1FastDecompressorCodec — Type

LZO1X1FastDecompressorCodec <: TranscodingStreams.Codec

A struct that decompresses data using liblzo1 library version of the LZO 1X1 algorithm.

The LZO 1X1 algorithm is a Lempel-Ziv lossless compression algorithm. Compressed streams consist of alternating encoded instructions and sequences of literal values. The encoded instructions tell the decompressor to either:

copy a sequence of bytes of a particular length directly from the input to the output (literal copy), or

look back a certain distance in the already returned output and copy a sequence of bytes of a particular length from the output to the output again.

The C implementation of LZO defined by liblzo2 requires that all decompressed information be available in working memory at once, and therefore does not take advantage of the memory savings allowed by TranscodingStreams. The C library version claims to use no additional working memory, but it requires that a continuous space in memory be available to hold the entire output of the compressed data, the length of which is not knowable a priori but can be larger than the compressed data by a factor of roughly 255. This implementation reports a very large memory requirement with TranscodingStreams.minoutsize to account for this.

Base.copyto! — Method

copyto!(dest::CircularVector, do, src::CircularVector, so, N)

Copy N elements from collection src starting at the linear index so to array dest starting at the index do. Return dest.

Because CircularVectors have circular boundary conditions on the indices, the linear indices do and so may be negative, zero, or greater than the last indices of dest and src, respective, and the number of elements copied N may be greater than the lengths of either.

Base.copyto! — Method

copyto!(dest::CircularVector, do, src::Memory, so, N)

Copy N elements from collection src starting at the linear index so to array dest starting at the index do. Return dest.

Because CircularVectors have circular boundary conditions on the indices, the linear index do may be negative, zero, or greater than lastindex(dest), and the number of elements copied N may be greater than length(dest).

Base.copyto! — Method

copyto!(dest::Memory, do, src::CircularVector, so, N)

Copy N elements from collection src starting at the linear index so to array dest starting at the index do. Return dest.

Because CircularVectors have circular boundary conditions on the indices, the linear index so may be negative, zero, or greater than lastindex(src), and the number of elements copied N may be greater than length(src).

Base.copyto! — Method

copyto!(dest::CircularVector{T}, do, src::AbstractVector{T}, so, N)

Copy N elements from collection src starting at the linear index so to array dest starting at the index do. Return dest.

Because CircularVectors have circular boundary conditions on the indices, the linear index do may be negative, zero, or greater than lastindex(dest), and the number of elements copied N may be greater than length(dest).

CodecLZO.command_length — Function

command_length(command, [last_literal_length=0])::Int

Return the number of bytes in the encoded commands.

CodecLZO.compute_run_remainder — Method

compute_run_remainder(n, bits)::Tuple{Int, Int}

Compute the number of bytes necessary to encode a run of length n given a first-byte mask of length bits, also returning the remainder byte.

Note

This method does not adjust n before computing the run length. Perform adjustments before calling this method.

CodecLZO.decode — Function

decode(::Type{CommandPair}, data, [start_index=1; first_literal=false, last_literal_length=0])::Tuple{Int, CommandPair}

Decode a CommandPair from a byte array data starting at start_index.

Returns a tuple of the number of bytes read from data and the decoded CommandPair.

Positional arguments

data: a linear-indexed collection of UInt8 values that has getindex(data, ::Integer) and lastindex(data) methods defined.

start_index::Integer = 1: where in data to begin decoding.

Keyword arguments

first_literal::Bool = false: if true, the decoding algorithm will ignore the history copy and use the special encoding for the first literal of the stream.

last_literal_length::Integer = 0: if 0-3, apply special decoding for the history copy to allow short copies.

CodecLZO.decode_run — Function

decode_run(input, bits::Integer, [start_index::Integer = 1])::Tuple{Int, Int}

Decode the number of bytes in the encoding and the length of the run in bytes of the run in input given a mask of bits bits, optionally starting decoding at index start_index.

CodecLZO.encode! — Function

encode!(data, cp::CommandPair, [start_index = 1; last_literal_length = 0])::Int

Encode CommandPair cp to byte array data.

Returns the number of bytes written to data or zero if the encoding failed.

Positional arguments

data: a linear-indexed collection of UInt8 values that has setindex!(data, ::Integer) and lastindex(data) methods defined.

start_index::Integer = 1: where in data to begin decoding.

Keyword arguments

last_literal_length::Integer = 0: if 0-3, apply special encoding for the history copy to allow short copies.

CodecLZO.encode_run! — Function

encode_run!(output, len, bits, [start_index=1])::Int

Emit the number of zero bytes necessary to encode a length len in a command expecting bits leading bits to output, optionally starting at index start_index, returning the number of bytes written.

Literal and history copy lengths are always encoded as either a single byte or a sequence of three or more bytes. If len < (1 << bits), the length will be encoded in the lower bits bits of the starting byte of output so the return will be 1. Otherwise, the return will be the number of bytes needed to encode the length.

If output is not large enough to hold the length of the run, this function returns 0 and output is unchanged.

Arguments:

output: The target array-like object that will be written to. The type only needs to implement setindex!(output, ::UInt8, ::Int) and lastindex(output).

len::Integer: The adjusted length to encode (see note).

bits::Integer: The number of bits that makes up the length mask of the command.

start_index::Integer=1: Where to begin writing the encoded run in output.

Note

The argument len is expected to be the adjusted length for the command. Literals use an adjusted length of len = length(literal) - 3 and copy commands use an adjusted literal length of len = length(copy) - 2.

CodecLZO.encode_run — Method

encode_run(len::Integer, bits::Integer)::Vector{UInt8}

Emit a vector of the number of zero bytes necessary to encode a length len in a command expecting bits leading bits.

See: encode_run!.

CodecLZO.lzo_compress — Function

lzo_compress(src, [working_memory=zeros(UInt8, 1<<12)])::Vector{UInt8}

Compress src using the LZO 1X1 algorithm.

Returns a compressed version of src.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_compress! — Function

lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])

Compress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the compressed data if necessary. Returns the modified dest.

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.lzo_decompress! — Method

lzo_decompress!(dest::Vector{UInt8}, src)

Decompress src to dest using the LZO 1X1 algorithm.

The destination vector dest will be resized to fit the decompressed data if necessary. Returns the modified dest.

CodecLZO.lzo_decompress — Method

lzo_decompress(src)::Vector{UInt8}

Decompress src using the LZO 1X1 algorithm.

Returns a decompressed version of src.

CodecLZO.match_length — Method

match_length(v, i::Integer, j::Integer, N::Integer)::Int

Searches v[i:i+N-1] for the last element that matches in v[j:j+N-1], returning the maximal length of the matching elements up to and including N.

The argument v need only implement getindex(v, ::Integer). If the search runs out of bounds, a BoundsError will be thrown. If no matches are found, 0 will be returned, and if no non-match is found, N will be returned.

CodecLZO.maxcopy! — Method

maxcopy!(dest, start_index, src)::Int

Copy as much of src into dest starting at index start_index as possible.

Requires lastindex(dest), length(src), and copyto!(dest, start_index, src, 1, ::Int) to be defined for dest and src types.

Returns the number of bytes copied.

CodecLZO.multiplicative_hash — Method

multiplicative_hash(value, magic_number, bits, [mask::V = typemax(UInt64)])

Hash value into a type V using multiplicative hashing.

This method performs floor((value * magic_number % W) / (W / M)) where W = 2^64, M = 2^m, and magic_number is relatively prime to W, is large, and has a good mix of 1s and 0s in its binary representation. In modulo 2^64 arithmetic, this becomes (value * magic_number) >>> m.

CodecLZO.reinterpret_get — Method

reinterpret_get(T::Type, input, [index::Integer = 1])::T

Reinterpret bytes from input as an LE-ordered value of type T, optionally starting at index. This tries to be faster than reinterpret(T, input[index:index+sizeof(T)-1]).

input can be anything that is linearly indexed and getindex(input, ::Int) returns a type S for which the operator |(::T, ::S) is defined.

CodecLZO.reinterpret_next — Method

reinterpret_next(previous::T, input::AbstractVector{UInt8}, [index::Int = 1])::T

Get the byte from input at index and push it to the LSB of previous, rotating off the MSB. This tries to be faster than doing reinterpret(T, input[index:index+sizeof(T)-1]) twice by reusing the already read LSBs.

CodecLZO.replace_all_matching! — Method

replace_all_matching!(h::HashMap, input, input_start, output, output_start)

Count the number of elements at the start of input that match the elements at the start of output, putting the matching indices of input as values into h keyed by the K integer read from input at that index.

Returns the number of matching bytes found (not necessarily equal to the number of Ks put into h).

CodecLZO.state — Method

state(codec)::MatchingState

Return the state of the compressor.

The state can be one of: - HISTORY: The codec is looking for the end of the historical match (i.e., the byte before the next literal). - LITERAL: The codec is looking for the next historical match (i.e., the length of the current literal). - COMMAND: The codec is writing a command sequence. - FLUSH: The codec is flushing the literal buffer to output.

CodecLZO.state — Method

state(codec)::MatchingState

Return the state of the decompressor.

The state can be one of: - AWAITING_FIRST_LITERAL: The codec is looking for the end of the first literal copy command. Transitions to READINGLITERAL. - `AWAITINGCOMMAND: The codec is looking for the next command pair. Transitions to COPYING_HISTORY. -COPYINGHISTORY: The codec is copying a sequence of bytes from history in the output buffer to the write head of the output buffer, evicting bytes from the back of the output buffer to output memory. Transitions to FLUSHING. -READINGLITERAL: The codec is writing a literal sequence to the output buffer. Transitions to FLUSHING. -FLUSHING: The codec is flushing the output buffer to output memory. Transitions to AWAITING_COMMAND, COPYING_HISTORY, or READING_LITERAL, depending on if the output memory blocked during a sequence write. -ENDOFSTREAM`: The codec received the end-of-stream command. Any further input will result in an exception being thrown.

CodecLZO.unsafe_lzo_compress! — Function

unsafe_lzo_compress!(dest::Vector{UInt8}, src, [working_memory=zeros(UInt8, 1<<12)])::Int

Compress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the compressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to compress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).

Pass working_memory, a Vector{UInt8} with length(working_memory) >= 1<<12, to reuse pre-allocated memory required by the algorithm.

CodecLZO.unsafe_lzo_decompress! — Method

unsafe_lzo_decompress!(dest::Vector{UInt8}, src)::Int

Decompress src to dest using the LZO 1X1 algorithm.

The method is "unsafe" in that it does not check to see if the decompressed output can fit into dest before proceeding, and may write out of bounds or crash your program if the number of bytes required to decompress src is larger than the number of bytes available in dest. The method returns the number of bytes written to dest, which may be greater than length(dest).